# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Balance:
    properties:
      id:
        type: integer
      time:
        type: string
      value:
        type: number
    type: object
  domain.Payout:
    properties:
      amount:
        type: number
      id:
        type: integer
      time:
        type: string
    type: object
...
  eliftech-school_backend_internal_controllers.balancesResponse:
    properties:
      balances:
        items:
          $ref: '#/definitions/domain.Balance'
        type: array
    type: object
  eliftech-school_backend_internal_controllers.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  eliftech-school_backend_internal_controllers.payoutsResponse:
    properties:
      payouts:
        items:
          $ref: '#/definitions/domain.Payout'
        type: array
    type: object
...
  internal_controllers.balancesResponse:
    properties:
      balances:
        items:
          $ref: '#/definitions/domain.Balance'
        type: array
    type: object
  internal_controllers.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_controllers.payoutsResponse:
    properties:
      payouts:
        items:
          $ref: '#/definitions/domain.Payout'
        type: array
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Back end for 50 Cent Application
  title: 50 Cent API
  version: "1.1"
paths:
...
  /consumers/balance:
    get:
      consumes:
      - application/json
      description: Get List of balance snapshots
      operationId: getBalanceHistoryConsumer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.balancesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Balance history for consumer
      tags:
      - Consumer
  /consumers/payments:
    get:
      consumes:
      - application/json
      description: Get List of required payments snapshots
      operationId: getRequiredPayments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.payoutsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get required payments for consumer
      tags:
      - Consumer
  /investors/balance:
    get:
      consumes:
      - application/json
      description: Get List of balance snapshots
      operationId: getBalanceHistoryInvestor
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.balancesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Balance history for investor
      tags:
      - Investor
  /investors/payouts:
    get:
      consumes:
      - application/json
      description: Get List of potential payouts snapshots
      operationId: getPotentialPayouts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.payoutsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get potential payouts for investor
      tags:
      - Investor
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"